Make Dialogs a service which can be injected in the view model
To make dialog-opening testable in the viewModel.
Suggested in http://stackoverflow.com/a/1044304/873661.
